cogl/egl: Use EGL_MESA_unsynchronized_image
This extension allows Mutter to inform the EGL driver that a texture
created from an EGLImage doesn't need to have its contents automatically
updated to correspond to the image. In some drivers (such as v3d) the
texture ends up using an internal copy and without this extension the
driver can't know when it needs to update that copy. With this extension
we can inform the driver when the underlying pixmap has received damage.